refactor: point 패키지 생성
- Point, File, Rank 포함
