+++ src/main/java/chess/controller/ChessController.java
@@ -3,9 +3,9 @@ package chess.controller;
 import chess.domain.BoardFactory;
 import chess.domain.ChessGame;
 import chess.domain.Command;
-import chess.domain.File;
-import chess.domain.Point;
-import chess.domain.Rank;
+import chess.domain.point.File;
+import chess.domain.point.Point;
+import chess.domain.point.Rank;
 import chess.view.InputView;
 import chess.view.OutputView;
