+++ src/main/java/chess/domain/ChessGame.java
@@ -4,6 +4,7 @@ import chess.domain.piece.Piece;
 import chess.domain.piece.Team;
 import chess.domain.player.Player;
 import chess.domain.player.Players;
+import chess.domain.point.Point;
 import java.util.Map;
 
 public class ChessGame {
