+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,6 +1,6 @@
 package chess.domain.piece;
 
-import chess.domain.Point;
+import chess.domain.point.Point;
 
 public abstract class Piece {
