+++ src/test/java/chess/domain/BoardTest.java
@@ -6,6 +6,9 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import chess.domain.piece.Piece;
 import chess.domain.piece.Team;
 import chess.domain.player.Player;
+import chess.domain.point.File;
+import chess.domain.point.Point;
+import chess.domain.point.Rank;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
