+++ src/test/java/chess/domain/piece/QueenTest.java
@@ -2,9 +2,9 @@ package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.File;
-import chess.domain.Point;
-import chess.domain.Rank;
+import chess.domain.point.File;
+import chess.domain.point.Point;
+import chess.domain.point.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
