+++ src/test/java/chess/domain/point/FileTest.java
@@ -1,8 +1,9 @@
-package chess.domain;
+package chess.domain.point;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import chess.domain.point.File;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
